0.3.2

+++ mocap/device_functions.py
@@ -28,8 +28,9 @@ def new_process_send_images(
     
     """ Send images to the server """
     
-    cv2.namedWindow("frame")
+    debug_flag = False
     
+    if debug_flag: cv2.namedWindow("frame")
     shared_np = np.frombuffer(shared_memory.get_obj(), dtype=np.uint8)
     pub_port = config["PUBLISHERS"]["PORT_IMG"][0]
     sender = ImageSender(connect_to=f"tcp://*:{pub_port}", REQ_REP=False)
@@ -40,16 +41,13 @@ def new_process_send_images(
         with shared_memory.get_lock():
             frame = np.copy(shared_np)
             frame = frame.reshape((im_height, im_width))
-        
-        cv2.imshow("frame", frame)
-        cv2.waitKey(1)
-        
+        if debug_flag:
+            cv2.imshow("frame", frame)
+            cv2.waitKey(1)
         _, frame = cv2.imencode(".jpg", frame, [int(cv2.IMWRITE_JPEG_QUALITY), 95])
         sender.send_jpg_pubsub(f"{msg}", frame)
     print(f"MocapCamera::SUBPROCESS::new_process_send_images: Stop process")
-    
-    cv2.destroyWindow("frame")
-    
+    if debug_flag: cv2.destroyWindow("frame")
     return
 
 
@@ -71,7 +69,6 @@ def new_process_send_detections(
     elif detection_method == "FOREGROUND":
         cMoG = cv2.createBackgroundSubtractorMOG2()
         detection_function = lambda frame: detect_by_foreground(frame, config["DETECTION"]["FOREGROUND"], cMoG)
-    
     shared_np = np.frombuffer(shared_memory.get_obj(), dtype=np.uint8)
     pub_port = config["PUBLISHERS"]["PORT_PTS"][0]
     sender = ImageSender(connect_to=f"tcp://*:{pub_port}", REQ_REP=False)
@@ -99,25 +96,42 @@ def new_process_send_all(
     
     """ Send both images and detections to the server """
     
-    # cv2.namedWindow("detection")
-    # cv2.namedWindow("debug")
+    foreground_detection_mode = 3
+    debug_flag = False
     
+    # imgSaver = ImageSaver(
+    #     config["DEVICE"]["DIR"],
+    #     config["DEVICE"]["NB"],
+    #     config["DEVICE"]["STEP"],
+    #     config["DEVICE"]["CLEAR"])
+    
+    if debug_flag:
+        cv2.namedWindow("detection")
+        cv2.namedWindow("debug")
     detection_method = config["DETECTION"].get("METHOD", "MARKERS")
     detection_function = None
     if detection_method == "MARKERS":
+        detection_mode = 1
         detection_function = lambda frame: detect_by_markers(frame, config["DETECTION"]["MARKERS"])
+        first_iter_flag = False
     elif detection_method == "FOREGROUND":
-        # machine = cv2.createBackgroundSubtractorMOG2()
-        # machine = cv2.createBackgroundSubtractorKNN()
-        background = np.zeros((1080, 1440), dtype=np.uint8)
-        foreground_mask = np.ones((1080, 1440), dtype=np.bool8)
-        first_iter_flag = True
-        detection_function = lambda frame: detect_by_foreground(frame, config["DETECTION"]["FOREGROUND"], 
-                                                                machine=None, 
+        detection_mode = 2
+        machine = None
+        background = None
+        foreground_mask = None
+        first_iter_flag = False
+        if foreground_detection_mode == 1: machine = cv2.createBackgroundSubtractorMOG2()
+        if foreground_detection_mode == 2: machine = cv2.createBackgroundSubtractorKNN()
+        if foreground_detection_mode == 3:
+            background = np.zeros((1080, 1440), dtype=np.uint8)
+            foreground_mask = np.ones((1080, 1440), dtype=np.bool8)
+            first_iter_flag = True
+        detection_function = lambda frame: detect_by_foreground(frame,
+                                                                config["DETECTION"]["FOREGROUND"], 
+                                                                machine=machine, 
                                                                 background=background, 
                                                                 foreground_mask=foreground_mask)
     detection_limit = config["DETECTION"].get("MAX_DETECTIONS", 10)
-    
     shared_np = np.frombuffer(shared_memory.get_obj(), dtype=np.uint8)
     pub_port_img = config["PUBLISHERS"]["PORT_IMG"][0]
     sender_img = ImageSender(connect_to=f"tcp://*:{pub_port_img}", REQ_REP=False)
@@ -130,16 +144,13 @@ def new_process_send_all(
         with shared_memory.get_lock():
             frame = np.copy(shared_np)
             frame = frame.reshape((im_height, im_width))
-        
-        if detection_method == "FOREGROUND" and first_iter_flag:
+        if first_iter_flag:
             background = frame.copy()
             first_iter_flag = False
-        
         objs = detection_function(frame)
         if len(objs) > detection_limit:
             objs = objs[:detection_limit]
-        
-        if detection_method == "FOREGROUND":
+        if detection_mode == 2:
             i = 1
             x_g = 0
             y_g = 0
@@ -150,20 +161,19 @@ def new_process_send_all(
             x_g = int(x_g / i)
             y_g = int(y_g / i)
             objs = [[x_g, y_g, 0]]
-        
-        # blank = np.concatenate([frame[:,:,None], frame[:,:,None], frame[:,:,None]], axis=-1)
-        # cv2.circle(blank, (x_g, y_g), int(4), (0,0,255), 6)
-        # cv2.imshow("detection", cv2.resize(blank, None, fx=0.5, fy=0.5))
-        # cv2.waitKey(1)
-        
+        if debug_flag:
+            blank = np.concatenate([frame[:,:,None], frame[:,:,None], frame[:,:,None]], axis=-1)
+            if detection_mode == 2: cv2.circle(blank, (x_g, y_g), int(4), (0,0,255), 6)
+            cv2.imshow("detection", cv2.resize(blank, None, fx=0.5, fy=0.5))
+            cv2.waitKey(1)
+        # imgSaver.save_image(blank)
         _, frame = cv2.imencode(".jpg", frame, [int(cv2.IMWRITE_JPEG_QUALITY), 95])
         sender_img.send_jpg_pubsub(f"{msg}", frame)
         sender_det.send_image_pubsub(f"{msg}", np.array(objs, dtype=np.float32))
-        
-    # cv2.destroyWindow("detection")
-    # cv2.destroyWindow("debug")
-        
     print(f"MocapCamera::SUBPROCESS::new_process_send_all: Stop process")
+    if debug_flag:
+        cv2.destroyWindow("detection")
+        cv2.destroyWindow("debug")
     return
 
 
@@ -295,6 +305,7 @@ def new_process_calculate_projection_matrix(
         with shared_memory.get_lock():
             frame = np.copy(shared_np)
             frame = frame.reshape((im_height, im_width))
+        
         objs = detect_by_markers(frame, config["DETECTION"]["MARKERS"])
         flag_en = (len(objs) == PATTERN_PNB)
         if flag_en:
@@ -303,12 +314,12 @@ def new_process_calculate_projection_matrix(
         blank = np.zeros((im_height, im_width, 3), dtype=np.uint8)
         for i in [0,1,2]: blank[:,:,i] = frame
         for x, y, r in objs:
-            cv2.circle(blank, (int(x), int(y)), int(r*2), (0,0,255), 2)
+            cv2.circle(blank, (int(x), int(y)), int(r*3), (0,0,255), 4)
         if flag_en:
             for idx, x, y in sorted_markers:
-                cv2.putText(blank, f"{idx+1}", (int(x)+10, int(y-10)), cv2.FONT_HERSHEY_SIMPLEX, 0.8, (0,0,255), 2)
+                cv2.putText(blank, f"{idx+1}", (int(x)+10, int(y-10)), cv2.FONT_HERSHEY_SIMPLEX, 0.8, (0,0,255), 4)
         blank = cv2.resize(blank, None, fx=scale_display, fy=scale_display)
-        cv2.putText(blank, f"DET STRUCT SHAPE: {np.array(objs).shape}", (10,10), cv2.FONT_HERSHEY_SIMPLEX, 0.8, (0,0,255), 2)
+        cv2.putText(blank, f"DET STRUCT SHAPE: {np.array(objs).shape}", (30,10), cv2.FONT_HERSHEY_SIMPLEX, 0.8, (0,0,255), 2)
         cv2.imshow("frame", blank)
         resp = cv2.waitKey(1)
         if resp == ord('s') and flag_en:
